LPD-23832 This logic is better suited here
