+++ modules/apps/company/company-test/src/testIntegration/java/com/liferay/company/service/test/CompanyLocalServiceDBPartitionTest.java
@@ -7,23 +7,15 @@ package com.liferay.company.service.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.counter.kernel.service.CounterLocalService;
-import com.liferay.petra.string.StringBundler;
-import com.liferay.petra.string.StringPool;
 import com.liferay.portal.db.partition.db.DBPartitionDB;
 import com.liferay.portal.db.partition.test.util.BaseDBPartitionTestCase;
 import com.liferay.portal.db.partition.util.DBPartitionUtil;
 import com.liferay.portal.kernel.dao.db.DBType;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.instance.PortalInstancePool;
-import com.liferay.portal.kernel.messaging.Message;
 import com.liferay.portal.kernel.model.Company;
 import com.liferay.portal.kernel.model.ResourceAction;
 import com.liferay.portal.kernel.model.VirtualHost;
-import com.liferay.portal.kernel.scheduler.SchedulerEngine;
-import com.liferay.portal.kernel.scheduler.StorageType;
-import com.liferay.portal.kernel.scheduler.TimeUnit;
-import com.liferay.portal.kernel.scheduler.Trigger;
-import com.liferay.portal.kernel.scheduler.TriggerFactory;
 import com.liferay.portal.kernel.service.ResourceActionLocalService;
 import com.liferay.portal.kernel.service.VirtualHostLocalService;
 import com.liferay.portal.kernel.test.ReflectionTestUtil;
@@ -45,7 +37,6 @@ import com.liferay.portal.test.rule.PermissionCheckerMethodTestRule;
 import com.liferay.portal.test.rule.TransactionalTestRule;
 
 import java.sql.DatabaseMetaData;
-import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
@@ -54,10 +45,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 
-import org.junit.After;
 import org.junit.AfterClass;
 import org.junit.Assert;
-import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
 import org.junit.Rule;
@@ -95,17 +84,6 @@ public class CompanyLocalServiceDBPartitionTest
 	@AfterClass
 	public static void tearDownClass() throws Exception {
 		_regenerateResourceActions();
-	}
-
-	@Before
-	public void setUp() throws Exception {
-		_scheduleJob(_defaultCompanyId, _JOB_NAME);
-		_scheduleJob(_defaultCompanyId, _JOB_NAME + "test");
-	}
-
-	@After
-	public void tearDown() throws Exception {
-		_schedulerEngine.delete(_JOB_GROUP_NAME, StorageType.PERSISTED);
 	}
 
 	@Test
@@ -215,14 +193,7 @@ public class CompanyLocalServiceDBPartitionTest
 	public void testAddDBPartitionCompany() throws Exception {
 		Company company = CompanyTestUtil.addCompany();
 
-		_scheduleJob(company.getCompanyId(), _JOB_NAME);
-
-		Assert.assertEquals(_JOBS_COUNT + 1, _getJobsCount(_defaultCompanyId));
-
 		companyLocalService.extractDBPartitionCompany(company.getCompanyId());
-
-		Assert.assertEquals(_JOBS_COUNT, _getJobsCount(_defaultCompanyId));
-		Assert.assertEquals(1, _getJobsCount(company.getCompanyId()));
 
 		String name = "new" + company.getName();
 		String virtualHostName = "new" + company.getVirtualHostname();
@@ -243,9 +214,6 @@ public class CompanyLocalServiceDBPartitionTest
 			Assert.assertEquals(name, company.getName());
 			Assert.assertEquals(virtualHostName, company.getVirtualHostname());
 			Assert.assertEquals(webId, company.getWebId());
-
-			Assert.assertEquals(
-				_JOBS_COUNT + 1, _getJobsCount(_defaultCompanyId));
 		}
 		finally {
 			if (standaloneDBPartition) {
@@ -263,10 +231,6 @@ public class CompanyLocalServiceDBPartitionTest
 
 		Company company = CompanyTestUtil.addCompany();
 
-		_scheduleJob(company.getCompanyId(), _JOB_NAME);
-
-		Assert.assertEquals(_JOBS_COUNT + 1, _getJobsCount(_defaultCompanyId));
-
 		boolean standaloneDBPartition = false;
 
 		try {
@@ -274,9 +238,6 @@ public class CompanyLocalServiceDBPartitionTest
 				company.getCompanyId());
 
 			standaloneDBPartition = true;
-
-			Assert.assertEquals(_JOBS_COUNT, _getJobsCount(_defaultCompanyId));
-			Assert.assertEquals(1, _getJobsCount(company.getCompanyId()));
 
 			Company defaultCompany = companyLocalService.getCompany(
 				_defaultCompanyId);
@@ -294,10 +255,6 @@ public class CompanyLocalServiceDBPartitionTest
 				Assert.assertFalse(
 					ArrayUtil.contains(
 						_getCompanyIdsBySQL(), company.getCompanyId()));
-
-				Assert.assertEquals(
-					_JOBS_COUNT, _getJobsCount(_defaultCompanyId));
-				Assert.assertEquals(1, _getJobsCount(company.getCompanyId()));
 
 				_checkStandaloneDBPartitionTables(
 					company.getCompanyId(), "Company", "VirtualHost");
@@ -319,10 +276,6 @@ public class CompanyLocalServiceDBPartitionTest
 
 		Company company = CompanyTestUtil.addCompany();
 
-		_scheduleJob(company.getCompanyId(), _JOB_NAME);
-
-		Assert.assertEquals(_JOBS_COUNT + 1, _getJobsCount(_defaultCompanyId));
-
 		boolean standaloneDBPartition = false;
 
 		try {
@@ -330,9 +283,6 @@ public class CompanyLocalServiceDBPartitionTest
 				company.getCompanyId());
 
 			standaloneDBPartition = true;
-
-			Assert.assertEquals(_JOBS_COUNT, _getJobsCount(_defaultCompanyId));
-			Assert.assertEquals(1, _getJobsCount(company.getCompanyId()));
 
 			try (AutoCloseable autoCloseable =
 					ReflectionTestUtil.setFieldValueWithAutoCloseable(
@@ -364,10 +314,6 @@ public class CompanyLocalServiceDBPartitionTest
 					ArrayUtil.contains(
 						_getCompanyIdsBySQL(), company.getCompanyId()));
 
-				Assert.assertEquals(
-					_JOBS_COUNT, _getJobsCount(_defaultCompanyId));
-				Assert.assertEquals(1, _getJobsCount(company.getCompanyId()));
-
 				_checkStandaloneDBPartitionTables(
 					company.getCompanyId(), "Company", "VirtualHost");
 			}
@@ -385,10 +331,6 @@ public class CompanyLocalServiceDBPartitionTest
 	@Test
 	public void testCopyDBPartitionCompany() throws Exception {
 		Company company = CompanyTestUtil.addCompany();
-
-		_scheduleJob(company.getCompanyId(), _JOB_NAME);
-
-		Assert.assertEquals(_JOBS_COUNT + 1, _getJobsCount(_defaultCompanyId));
 
 		String name = RandomTestUtil.randomString();
 		String virtualHostname = RandomTestUtil.randomString();
@@ -408,11 +350,6 @@ public class CompanyLocalServiceDBPartitionTest
 			Assert.assertEquals(
 				virtualHostname, newCompany.getVirtualHostname());
 			Assert.assertEquals(webId, newCompany.getWebId());
-
-			Assert.assertEquals(
-				_JOBS_COUNT + 2, _getJobsCount(_defaultCompanyId));
-			Assert.assertEquals(
-				1, _getCompanyJobsCount(company.getCompanyId()));
 		}
 		finally {
 			removeDBPartitions(new long[] {company.getCompanyId()});
@@ -429,10 +366,6 @@ public class CompanyLocalServiceDBPartitionTest
 
 		Company company = CompanyTestUtil.addCompany();
 
-		_scheduleJob(company.getCompanyId(), _JOB_NAME);
-
-		Assert.assertEquals(_JOBS_COUNT + 1, _getJobsCount(_defaultCompanyId));
-
 		long toCompanyId = RandomTestUtil.nextLong();
 
 		try {
@@ -446,9 +379,6 @@ public class CompanyLocalServiceDBPartitionTest
 			Assert.assertFalse(
 				ArrayUtil.contains(_getCompanyIdsBySQL(), toCompanyId));
 
-			Assert.assertEquals(
-				_JOBS_COUNT + 1, _getJobsCount(_defaultCompanyId));
-
 			_checkPartitionNonexists(toCompanyId);
 		}
 		finally {
@@ -461,10 +391,6 @@ public class CompanyLocalServiceDBPartitionTest
 		throws Exception {
 
 		Company company = CompanyTestUtil.addCompany();
-
-		_scheduleJob(company.getCompanyId(), _JOB_NAME);
-
-		Assert.assertEquals(_JOBS_COUNT + 1, _getJobsCount(_defaultCompanyId));
 
 		String name = RandomTestUtil.randomString();
 		long toCompanyId = RandomTestUtil.nextLong();
@@ -499,9 +425,6 @@ public class CompanyLocalServiceDBPartitionTest
 			Assert.assertFalse(
 				ArrayUtil.contains(_getCompanyIdsBySQL(), toCompanyId));
 
-			Assert.assertEquals(
-				_JOBS_COUNT + 1, _getJobsCount(_defaultCompanyId));
-
 			_checkPartitionNonexists(toCompanyId);
 		}
 		finally {
@@ -513,10 +436,6 @@ public class CompanyLocalServiceDBPartitionTest
 	public void testDeleteCompany() throws Exception {
 		Company company = CompanyTestUtil.addCompany();
 
-		_scheduleJob(company.getCompanyId(), _JOB_NAME);
-
-		Assert.assertEquals(_JOBS_COUNT + 1, _getJobsCount(_defaultCompanyId));
-
 		int dbPartitionsCount = _getDBPartitionsCount();
 
 		companyLocalService.deleteCompany(company);
@@ -524,16 +443,11 @@ public class CompanyLocalServiceDBPartitionTest
 		Assert.assertFalse(
 			ArrayUtil.contains(_getCompanyIdsBySQL(), company.getCompanyId()));
 		Assert.assertEquals(dbPartitionsCount - 1, _getDBPartitionsCount());
-		Assert.assertEquals(_JOBS_COUNT, _getJobsCount(_defaultCompanyId));
 	}
 
 	@Test
 	public void testDeleteCompanyWhenDBPartitionUtilFails() throws Exception {
 		_company1 = CompanyTestUtil.addCompany();
-
-		_scheduleJob(_company1.getCompanyId(), _JOB_NAME);
-
-		Assert.assertEquals(_JOBS_COUNT + 1, _getJobsCount(_defaultCompanyId));
 
 		try (AutoCloseable autoCloseable =
 				ReflectionTestUtil.setFieldValueWithAutoCloseable(
@@ -559,18 +473,12 @@ public class CompanyLocalServiceDBPartitionTest
 			Assert.assertTrue(
 				ArrayUtil.contains(
 					_getCompanyIdsBySQL(), _company1.getCompanyId()));
-			Assert.assertEquals(
-				_JOBS_COUNT + 1, _getJobsCount(_defaultCompanyId));
 		}
 	}
 
 	@Test
 	public void testExtractDBPartitionCompany() throws Exception {
 		Company company = CompanyTestUtil.addCompany();
-
-		_scheduleJob(company.getCompanyId(), _JOB_NAME);
-
-		Assert.assertEquals(_JOBS_COUNT + 1, _getJobsCount(_defaultCompanyId));
 
 		boolean standaloneDBPartition = false;
 
@@ -583,9 +491,6 @@ public class CompanyLocalServiceDBPartitionTest
 					_getCompanyIdsBySQL(), company.getCompanyId()));
 
 			standaloneDBPartition = true;
-
-			Assert.assertEquals(_JOBS_COUNT, _getJobsCount(_defaultCompanyId));
-			Assert.assertEquals(1, _getJobsCount(company.getCompanyId()));
 
 			_checkStandaloneDBPartitionTables(
 				company.getCompanyId(), "Company", "VirtualHost");
@@ -605,10 +510,6 @@ public class CompanyLocalServiceDBPartitionTest
 		throws Exception {
 
 		Company company = CompanyTestUtil.addCompany();
-
-		_scheduleJob(company.getCompanyId(), _JOB_NAME);
-
-		Assert.assertEquals(_JOBS_COUNT + 1, _getJobsCount(_defaultCompanyId));
 
 		int tablesCount = _getTablesCount(company.getCompanyId());
 		int viewsCount = _getViewsCount(company.getCompanyId());
@@ -648,8 +549,6 @@ public class CompanyLocalServiceDBPartitionTest
 			Assert.assertTrue(
 				ArrayUtil.contains(
 					_getCompanyIdsBySQL(), company.getCompanyId()));
-			Assert.assertEquals(
-				_JOBS_COUNT + 1, _getJobsCount(_defaultCompanyId));
 		}
 		finally {
 			if (standaloneDBPartition) {
@@ -724,22 +623,6 @@ public class CompanyLocalServiceDBPartitionTest
 			PortalInstancePool.class, "_getCompanyIdsBySQL", null, null);
 	}
 
-	private int _getCompanyJobsCount(long companyId) throws Exception {
-		try (PreparedStatement preparedStatement = connection.prepareStatement(
-				StringBundler.concat(
-					"select count(1) from ", getPartitionName(companyId),
-					".QUARTZ_JOB_DETAILS where JOB_GROUP = '", _JOB_GROUP_NAME,
-					"' and JOB_NAME like '%@", companyId, "'"));
-			ResultSet resultSet = preparedStatement.executeQuery()) {
-
-			if (resultSet.next()) {
-				return resultSet.getInt(1);
-			}
-		}
-
-		throw new Exception("Table does not exist");
-	}
-
 	private int _getDBPartitionsCount() throws SQLException {
 		DatabaseMetaData databaseMetaData = connection.getMetaData();
 
@@ -756,24 +639,6 @@ public class CompanyLocalServiceDBPartitionTest
 		}
 
 		throw new SQLException("At least one database partition is required");
-	}
-
-	private int _getJobsCount(long companyId) throws Exception {
-		String partitionName = getPartitionName(companyId);
-
-		try (PreparedStatement preparedStatement = connection.prepareStatement(
-				StringBundler.concat(
-					"select count(1) from ", partitionName,
-					".QUARTZ_JOB_DETAILS where JOB_GROUP = '", _JOB_GROUP_NAME,
-					"'"));
-			ResultSet resultSet = preparedStatement.executeQuery()) {
-
-			if (resultSet.next()) {
-				return resultSet.getInt(1);
-			}
-		}
-
-		throw new Exception("Table does not exist");
 	}
 
 	private List<String> _getObjectNames(String objectType, long companyId)
@@ -810,22 +675,6 @@ public class CompanyLocalServiceDBPartitionTest
 		return viewNames.size();
 	}
 
-	private void _scheduleJob(long companyId, String jobName) throws Exception {
-		Trigger trigger = _triggerFactory.createTrigger(
-			StringBundler.concat(jobName, StringPool.AT, companyId),
-			_JOB_GROUP_NAME, null, null, 1, TimeUnit.DAY);
-
-		_schedulerEngine.schedule(
-			trigger, StringPool.BLANK, _JOB_GROUP_NAME, new Message(),
-			StorageType.PERSISTED);
-	}
-
-	private static final String _JOB_GROUP_NAME = "liferay/test";
-
-	private static final String _JOB_NAME = "test";
-
-	private static final int _JOBS_COUNT = 2;
-
 	@Inject
 	private static CounterLocalService _counterLocalService;
 
@@ -835,16 +684,6 @@ public class CompanyLocalServiceDBPartitionTest
 	private static ResourceActionLocalService _resourceActionLocalService;
 
 	private static Map<String, ResourceAction> _resourceActions;
-
-	@Inject(
-		filter = "component.name=com.liferay.portal.scheduler.quartz.internal.QuartzSchedulerEngine"
-	)
-	private static SchedulerEngine _schedulerEngine;
-
-	@Inject(
-		filter = "component.name=com.liferay.portal.scheduler.quartz.internal.QuartzTriggerFactory"
-	)
-	private static TriggerFactory _triggerFactory;
 
 	@Inject
 	private static VirtualHostLocalService _virtualHostLocalService;
